+++ zcash_client_backend/CHANGELOG.md
@@ -64,6 +64,9 @@ and this library adheres to Rust's notion of
     - `get_account_for_ufvk` now returns `Self::Account` instead of a bare
       `AccountId`.
     - Added `get_orchard_nullifiers` method.
+    - `get_transaction` now returns `Result<Option<Transaction>, _>` rather
+      than returning an `Err` if the `txid` parameter does not correspond to
+      a transaction in the database.
   - Changes to the `InputSource` trait:
     - `select_spendable_notes` now takes its `target_value` argument as a
       `NonNegativeAmount`. Also, it now returns a `SpendableNotes` data